Backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==4.2.16
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-crontab==0.7.1
django-environ==0.11.2
django-jalali==7.1.0
django-rest-knox==5.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
jalali_core==1.0.0
jdatetime==5.0.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
python-slugify==8.0.4
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
